requirements.txt
deutschland
pandas
numpy
os
customtkinter
matplotlib.pyplot
datetime
matplotlib.date
